Added ThemeColorView and replaced ColorInfoView
